Allow db_bind() to pass row_factory parameter to SQLite provider.
The sqlite3 driver allows for a row_factory parameter to connections
that lets one set a factory function for result rows.  This change
allows that parameter to be set when binding to a database so that one
can, for instance, get namedtuples back instead of tuples with no field
names.