Mention that PonyORM works with MariaDB as well
Because MySQL and MariaDB use the same on-the-wire protocol and most
SQL-level features are identical, PonyORM also works with MariaDB, using the
MySQL provider.

Tested on MariaDB 10.6.14.

After creating a 'test' database, and a user 'whoever' with access to that
database (`GRANT ALL PRIVILEGES ON test.* TO whoever`), the "first steps"
from https://docs.ponyorm.org/firststeps.html work fine, e.g.:

    from pony.orm import *
    db=Database()
    class Person(db.Entity):
        name=Required(str)
        age=Required(int)
    db.bind(provider='mysql', user='whoever', password='foobar', db='test')
    db.generate_mapping(create_tables=True)